+++ R/synthesis.R
@@ -34,6 +34,7 @@
 #' # Retrieve the graph signal using the synthesis operator
 #' f_rec <- synthesis(coef, tf)
 #' }
+#' @seealso \code{\link{analysis}}, \code{\link{tight_frame}}
 
 synthesis <- function(coeff, tf) {
   y <- t(coeff) %*% tf