+++ NEWS.md
@@ -17,9 +17,13 @@ gasper News
 -   add custom frame filters support (in `forward/inverse_sgwt`,
     `HPFVN`, `plot_filter` and `tight_frame`) by passing additional
     arguments (`filter_params`/`filter_func`).
--   add `graph_info` to retrieve from SuiteSparse the two/three tables
-    with “Matrix Information,” “Matrix Properties” and, if available,
-    “SVD Statistics”.
+-   add `get_graph_info` to retrieve from SuiteSparse the two/three
+    tables with “Matrix Information,” “Matrix Properties” and, if
+    available, “SVD Statistics”.
+-   change `info` output in `dowload_graph` to `temp` to access temp dir
+    instead.
+-   add optional logical argument `svd` to fetch singular values is
+    needed and provided.
 
 # gasper 1.1.4 relase (08/26/2023)